requirements.txt
Flask
python-dotenv
google-generativeai
tavily-python
trafilatura
readability-lxml
pypdf
requests
Werkzeug